proposal(select): avoid excessive getBoundingClientRect calls
Currently `md-select` passes the width of it's trigger element to it's overlay via a data binding. The data binding gets updated every time the change detection runs, which means that it'll run `getBoundingClientRect` on pretty much any interaction with the page. This could cause performance issues in the future, especially on slower device and more complex pages since every call to `getBoundingClientRect` causes a reflow.

This is a proposal to add a `match-origin` option to the `connected-overlay` directive. This option will make the overlay match it's origin's width/height and will only call `getBoundingClientRect` once when setting up the overlay.

The problem can alternatively be worked around with the following, however both approaches have their issues:
* Change the binding to `paneOpen ? _getWidth() : 0` - This works and avoids most of the `getBoundingClientRect` calls, but it looks weird and `getBoundingClientRect` for every change detection while the select is open.
* Pass in an observable to the `width` binding - This also works, but it ends up being inconvenient to have to deal with a type of `number | string | Observable` in the overlay directive.